requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.11
inflection==0.5.1
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
